+++ src/Phormium/Model.php
@@ -66,44 +66,97 @@ abstract class Model
         return $instance;
     }
 
+
     /**
-     * Fetches a single record by primary key.
+     * Fetches a single record by primary key, throws an exception if the model
+     * is not found. This method requires the model to have a PK defined.
      *
-     * @param mixed The primary key, can be more than 1 param for composite keys.
+     * @param mixed The primary key value, either as one or several arguments,
+     *      or as an array of one or several values.
+     * @return Model
      */
     public static function get()
     {
-        $args = func_get_args();
-        $count = func_num_args();
+        $argv = func_get_args();
+        $argc = func_num_args();
 
-        // Allow passing the PK as an array
-        if ($count == 1 && is_array($args[0])) {
-            $args = $args[0];
+        $qs = self::getQuerySetForPK($argv, $argc);
+        $model = $qs->single(true);
+
+        if ($model === null) {
+            $class = get_called_class();
+            $pk = implode(',', $argv);
+            throw new \Exception("[$class] record with primary key [$pk] does not exist.");
         }
 
-        $meta = self::getMeta();
+        return $model;
+    }
 
+    /**
+     * Fetches a single record by primary key, returns NULL if the model is not
+     * found. This method requires the model to have a PK defined.
+     *
+     * @param mixed The primary key value, either as one or several arguments,
+     *      or as an array of one or several values.
+     * @return Model|null The Model instance or NULL if not found.
+     */
+    public static function find()
+    {
+        $argv = func_get_args();
+        $argc = func_num_args();
+
+        $qs = self::getQuerySetForPK($argv, $argc);
+        return $qs->single(true);
+    }
+
+    /**
+     * Checks whether a record with the given Primary Key exists in the
+     * database. This method requires the model to have a PK defined.
+     *
+     * @param mixed The primary key value, either as one or several arguments,
+     *      or as an array of one or several values.
+     * @return boolean
+     */
+    public static function exists()
+    {
+        $argv = func_get_args();
+        $argc = func_num_args();
+
+        $qs = self::getQuerySetForPK($argv, $argc);
+        return $qs->exists();
+    }
+
+    /** Inner method used by get(), search() and exists(). */
+    private static function getQuerySetForPK($argv, $argc)
+    {
+        // Allow passing the PK as an array
+        if ($argc == 1 && is_array($argv[0])) {
+            $argv = $argv[0];
+            $argc = count($argv);
+        }
+
+        // Model must have PK defined
+        $meta = self::getMeta();
         if (!isset($meta->pk)) {
             $class = get_called_class();
             throw new \Exception("Primary key not defined for model [$class].");
         }
 
         // Check correct number of columns is given
-        $countArgs = count($args);
         $countPK = count($meta->pk);
-        if ($countArgs  !== $countPK) {
+        if ($argc !== $countPK) {
             $class = get_called_class();
-            throw new \Exception("Model [$class] has $countPK primary key columns. $countArgs arguments given.");
+            throw new \Exception("Model [$class] has $countPK primary key columns. $argc arguments given.");
         }
 
         // Create a queryset and filter by PK
         $qs = self::objects();
         foreach ($meta->pk as $name) {
-            $value = array_shift($args);
+            $value = array_shift($argv);
             $qs = $qs->filter($name, '=', $value);
         }
 
-        return $qs->single();
+        return $qs;
     }
 
     /**